backend/pyproject.toml
[project]
name = "ai-powered-form-filling"
version = "0.1.0"
description = "AI-Powered Form Filling Backend with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "fasttext>=0.9.3",
    "numpy<2",
    "paddleocr>=3.2.0",
    "paddlepaddle>=2.5.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    # Storage
    "minio>=7.2.0",
    # Configuration
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]
